change version to "0.8.3.a1" Better version comparison code that checks for correct django version moved new functions to utils.py to reduce imports which was causing problems when exporting functions.
