+++ drf_integrations/__init__.py
@@ -1,6 +1,9 @@
-from fields import get_json_model_field, get_json_model_field_import_name  # noqa
-from forms import get_json_form_field, get_json_model_field_import_name  # noqa
+from .utils import (  # noqa
+    get_json_form_field,
+    get_json_model_field,
+    get_json_model_field_import_name,
+)
 
-VERSION = "0.6.1"
+VERSION = "0.8.3.a1"
 
 default_app_config = "drf_integrations.apps.DRFIntegrationsConfig"
