+++ drf_integrations/fields.py
@@ -1,43 +1,12 @@
 import csv
-import django
 import io
-import logging
 from django import forms
-from django.conf import settings
 from django.core.exceptions import ValidationError
 from django.db import models
 from django.utils.encoding import force_str
-from django.utils.module_loading import import_string
 from django.utils.translation import ugettext_lazy as _
 
 from drf_integrations.utils import split_string
-
-logger = logging.getLogger(__name__)
-
-
-def get_json_model_field_import_name():
-    # if the Django version >= 3.2 then user the new default JSONField
-    if django.VERSION[0] >= 3 and django.VERSION[1] >= 2:
-        default = "django.db.models.JSONField"
-    else:
-        default = "django.contrib.postgres.fields.JSONField"
-
-    # Allow for override
-    model_field = getattr(settings, "DB_BACKEND_JSON_FIELD", default)
-    logger.info(f"Using django model JSONField: {model_field}")
-    return model_field
-
-
-def get_json_model_field():
-    model_field = get_json_model_field_import_name()
-    try:
-        return import_string(model_field)
-    except ImportError:
-        raise ImportError(
-            "drf_integrations can only work with a backend that supports JSON fields, "
-            "please make sure you set the DB_BACKEND_JSON_FIELD setting to the "
-            "JSONField of your backend."
-        )
 
 
 class CommaSeparatedValueField(models.TextField):
