+++ pyproject.toml
@@ -5,7 +5,7 @@
 
 [tool.poetry]
 name = "drf-integrations-framework"
-version = "0.8.2"
+version = "0.8.3.a1"
 description = "Django REST Framework plugin that simplifies the management of third party integrations"
 authors = ["Yoyo <dev@yoyowallet.com>"]
 license = "MIT"
